Refactored ValidationException handling
